// src/types/data/dailyRecordAPIData.type.ts
type DailyRecordAPIData = {
  daily_record_id: number;
  character_id: number;
  character_solved: number;
  weapon_id: number;
  weapon_solved: number;
  talent_id: number;
  talent_solved: number;
  constellation_id: number;
  constellation_solved: number;
  food_id: number;
  food_solved: number;
};

export default DailyRecordAPIData;